+++ examples/Reference/ReferenceServer/ReferenceServerConfiguration.cs
@@ -1,0 +1,84 @@
+#region Copyright (c) 2011-2021 Technosoftware GmbH. All rights reserved
+//-----------------------------------------------------------------------------
+// Copyright (c) 2011-2021 Technosoftware GmbH. All rights reserved
+// Web: https://technosoftware.com  
+// 
+// License: 
+// 
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+
+// The above copyright notice and this permission notice shall be included in all
+// copies or substantial portions of the Software.
+
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+// SOFTWARE.
+//
+// SPDX-License-Identifier: MIT
+//-----------------------------------------------------------------------------
+#endregion Copyright (c) 2011-2021 Technosoftware GmbH. All rights reserved
+
+#region Using Directives
+using System.Runtime.Serialization;
+#endregion
+
+namespace Technosoftware.ReferenceServer
+{
+    /// <summary>
+    /// Stores the configuration the Reference server.
+    /// </summary>
+    [DataContract(Namespace = Namespaces.Reference)]
+    public class ReferenceServerConfiguration
+    {
+        #region Constructors
+        /// <summary>
+        /// The default constructor.
+        /// </summary>
+        public ReferenceServerConfiguration()
+        {
+            Initialize();
+        }
+
+        /// <summary>
+        /// Initializes the object during deserialization.
+        /// </summary>
+        [OnDeserializing()]
+        private void Initialize(StreamingContext context)
+        {
+            Initialize();
+        }
+
+        /// <summary>
+        /// Sets private members to default values.
+        /// </summary>
+        private void Initialize()
+        {
+        }
+        #endregion
+
+        #region Public Properties
+        /// <summary>
+        /// Whether the user dialog for accepting invalid certificates should be displayed.
+        /// </summary>
+        [DataMember(Order = 1)]
+        public bool ShowCertificateValidationDialog
+        {
+            get { return m_showCertificateValidationDialog; }
+            set { m_showCertificateValidationDialog = value; }
+        }
+        #endregion
+
+        #region Private Members
+        private bool m_showCertificateValidationDialog;
+        #endregion
+    }
+}